app/requirements.txt
blinker==1.8.2
bson==0.5.10
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
flask==3.0.3
Flask-PyMongo==2.3.0
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
werkzeug==3.0.4
zipp==3.20.2
